BDB 6.0.19

+++ lang/sql/adapter/db_sequence.c
@@ -1,7 +1,7 @@
 /*
  * See the file LICENSE for redistribution information.
  *
- * Copyright (c) 2010, 2012 Oracle and/or its affiliates.  All rights reserved.
+ * Copyright (c) 2010, 2013 Oracle and/or its affiliates.  All rights reserved.
  *
  * $Id$
  */
@@ -160,7 +160,7 @@ static void db_seq_create_func(
 				    "%sInvalid parameter.", MSG_CREATE_FAIL);
 				goto err;
 			}
-			cookie.cache = sqlite3_value_int(argv[i]);
+			cookie.cache = (u32)sqlite3_value_int(argv[i]);			
 		} else if (strncmp((char *)sqlite3_value_text(argv[i]),
 		    "incr", 4) == 0) {
 			if (i == argc ||
@@ -322,18 +322,19 @@ static void db_seq_drop_func(
 		goto done;
 	}
 
+	/*
+	 * Drop the mutex - it's not valid to begin a transaction while
+	 * holding the mutex. We can drop it safely because it's use is to
+	 * protect handle cache changes.
+	 */
 	sqlite3_mutex_leave(pBt->mutex);
+
 	if ((rc = btreeSeqStartTransaction(context, p, 1)) != SQLITE_OK) {
 			btreeSeqError(context, SQLITE_ERROR,
 			    "Could not begin transaction for drop.");
 			return;
 	}
 
-	/*
-	 * Drop the mutex - it's not valid to begin a transaction while
-	 * holding the mutex. We can drop it safely because it's use is to
-	 * protect handle cache changes.
-	 */
 	sqlite3_mutex_enter(pBt->mutex);
 	btreeSeqRemoveHandle(context, p, cache_entry);
 done:	sqlite3_mutex_leave(pBt->mutex);
@@ -448,7 +449,7 @@ static void btreeSeqGetVal(
 		}
 		/* Cached gets can't be transactionally protected. */
 		if ((ret = cookie.handle->get(cookie.handle, NULL,
-		    cookie.incr, &val, 0)) != 0) {
+		    (u_int32_t)cookie.incr, &val, 0)) != 0) {
 			if (ret == EINVAL)
 				btreeSeqError(context, SQLITE_ERROR,
 				    "Sequence value out of bounds.");
@@ -883,7 +884,7 @@ static int btreeSeqPutCookie(
  * that case (and this function should not be called).
  *
  * It's safe to call this method multiple times since both
- * sqlite3BtreeBeginTrans and sqlite3BtreeBeginStmt are no-ops on subsequent
+ * btreeBeginTransInternal and sqlite3BtreeBeginStmt are no-ops on subsequent
  * calls.
  */
 static int btreeSeqStartTransaction(
